refactor(tui): extract shared render_entry_line into sprint_assign_list
Both single-card and multi-card sprint-assign renders now call a single
render_entry_line helper hosted in the sprint_assign_list module.
Eliminates the near-identical match-on-entry that lived in both
SprintAssignDialog::render and render_assign_multiple_cards_popup, so
future styling tweaks land in one place. Multi-card callers pass
current_sprint_id = None since that variant doesn't track a current
sprint.
